pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "schemaflux"
version = "1.0.0"
authors = [
    { name = "SchemaFlux Team" }
]
description = "A lightweight PostgreSQL migration library with performance analytics"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.0.0",
    "psycopg2-binary>=2.9.0",
    "pymongo>=4.10.1",
    "setuptools>=75.3.0",
    "wheel>=0.44.0",
]

[project.scripts]
schemaflux = "schemaflux.cli:cli"

[tool.setuptools]
packages = ["schemaflux"]
